Add a bundle parameter for PropertyStyler apply methods / applicators, and a bundle property to the Stylesheet protocol
The goal is to allow Stylesheets to be linked to assets in their own bundle.  So after creating a Stylesheet, it can be distributed in a framework or library that also includes an asset catalog with named images, colors, etc. and when property stylers go to apply those assets to a target, they will have the right bundle to pull them from.